pyproject.toml
[project]
name = "embodied-ai-toolkit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "rerun-sdk>=0.24.1",
    "robot-descriptions>=1.20.0",
    "tyro>=0.9.28",
    "viser>=1.0.6",
    "yourdfpy>=0.0.58",
]
